(refactor)(bash): embed the litmus job dependencies into test scripts ----------------------------------------------------------------------
The util invocation in 'before_script' needs kubeconfig. Moreover, gitlab
is inconsistent in the way 'before_script' & 'after_script' are treated, i.e.,
former is executed as part of same pod as the main (test) runner, while latter
is run in a separate pod. This makes .gitlab-ci.yml ungainly and difficult to
read. Therefore, embedded the litmus job dependency for a test runner in the runner's
script itself.

Signed-off-by: ksatchit <karthik.s@openebs.io>